feat: Support DICE authentication
With the latest changes, Linux-Akri may authenticate itself before
updating. To initiate the process, the `docker-image` attribute of
the POST request JSON must be equal to an empty string (""). Other-
wise, Linux-Akri will (as usual) update with the received image ar-
gument.

However, in case of an empty `docker-image` it is required that the
user provides an IP argument. For example, a possible `curl` command
to trigger a secure update is the following:

```
curl -H 'Content-Type: application/json' \
  -d '{ "args": ["--ip", "localhost"], "docker_image" : "" }' \
  -X POST localhost:8888/update
```

Finally, the `feat_linux_akri` branch of `ota-agent` - https://github.com/nubificus/ota-agent
has been prepared to support Linux-Akri updates. However, instead of the firmware path, now
we have to set a container image to be sent:
```
export CONTAINER_IMG=hello-world
```

Of course a Dice-Auth Attestation Server is required to run the entire process.
See https://github.com/nubificus/dice-auth

Signed-off-by: Ilias Lagomatis <ilago@nubificus.co.uk>